+++ drep/d_filter.py
@@ -282,7 +282,7 @@ def run_checkM(genome_folder,checkm_outf,**kwargs):
         chdb = pd.read_table(desired_file,sep='\t')
     except:
         logging.error("!!! checkM failed !!!\nIf using pyenv, make sure both python2 and " +\
-            "python3 are available (for example: pyenv pyenv global 3.5.1 2.7.9)")
+            "python3 are available (for example: pyenv global 3.5.1 2.7.9)")
         sys.exit()
     return chdb