+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [0.5.0] - 2017-03-16
+### Added
+- The final analyze plot can now be made!
+
+### Fixed
+- pyenv message
+- log typos
+
 ## [0.4.0] - 2017-03-10
 ### Fixed
 - prodigal now multithreads the correct amount